creating breadcrumbs functional

+++ app/Services/ProductService.php
@@ -4,24 +4,38 @@ namespace App\Services;
 
 use App\Models\Product;
 use Illuminate\Support\Arr;
+use Illuminate\Support\Facades\DB;
 
 class ProductService
 {
     public static function store(array $data): Product
     {
-        $product = Product::query()->create(Arr::get($data, 'product'));
-        ProductService::storeImages($product, Arr::get($data, 'images'));
-        self::storeProductParam($product, Arr::get($data, 'params'));
-        return $product;
+        try {
+            DB::beginTransaction();
+            $product = Product::query()->create(Arr::get($data, 'product'));
+            ProductService::storeImages($product, Arr::get($data, 'images'));
+            self::storeProductParam($product, Arr::get($data, 'params'));
+            DB::commit();
+            return $product;
+        } catch (\Exception $e) {
+            DB::rollBack();
+            abort(500, 'Store transaction failed');
+        }
     }
 
     public static function update(Product $product, array $data): Product
     {
-        ProductService::storeImages($product, Arr::get($data, 'images'));
-        ProductService::syncBatchParams($product, $data);
-        $product->update(Arr::get($data, 'product'));
-
-        return $product->fresh();
+        try {
+            DB::beginTransaction();
+            ProductService::storeImages($product, Arr::get($data, 'images'));
+            ProductService::syncBatchParams($product, $data);
+            $product->update(Arr::get($data, 'product'));
+            DB::commit();
+            return $product->fresh();
+        } catch (\Exception $e) {
+            DB::rollBack();
+            abort(500, 'Updated transaction failed');
+        }
     }
 
     public static function storeProductParam(Product $product, ?array $data = null): void
@@ -35,10 +49,13 @@ class ProductService
         }
     }
 
-    public static function syncBatchParams(Product $product, array $data): void
+    public static function syncBatchParams(Product $product, ?array $data = null): void
     {
-        $product->params()->detach();
-        ParamService::attachBatchParams($product, $data['params']);
+        if (isset($data['params']))
+        {
+            $product->params()->detach();
+            ParamService::attachBatchParams($product, $data['params']);
+        }
     }
 
     private static function storeImages(Product $product, ?array $images = null): void
